#349 rename package "org.xhtmlrenderer.simple" in module "flaying-saucer-swt"
to avoid the collision when the same package is found in two modules ("flaying-saucer-core" and "flaying-saucer-swt").
Java 9 modules doesn't allow such a collision.
